refactor: add validation for all DTOs

+++ src/main/java/academy/mediasoft/team/guideproject/entity/Rating.java
@@ -20,7 +20,7 @@ public class Rating {
     private Long id;
 
     @Column(nullable = false, name = "grade")
-    private Double grade;
+    private Integer grade;
 
     @Column(name = "created_at", nullable = false)
     @CreationTimestamp
